requirements.txt
celery
clickhouse-driver
Django
django-clickhouse
djangorestframework
djoser
psycopg2-binary
redis
requests